+++ addon/utils/reload-local-records.js
@@ -65,11 +65,10 @@ export function createLocalRecord(store, localAdapter, localStore, modelType, re
 
   dexieService.set('queueSyncDownWorksCount', dexieService.get('queueSyncDownWorksCount') + 1);
   if (record.get('id')) {
-    // TODO: if record from online store should not be unloaded then it will have dirty attribute.
-    record.set('syncDownTime', new Date());
     var snapshot = record._createSnapshot();
     let fieldsToUpdate = projection ? projection.attributes : null;
-    return new Ember.RSVP.Promise((resolve, reject) => localAdapter.addHashForBulkUpdateOrCreate(localStore, modelType, snapshot, fieldsToUpdate).then(() => {
+    return new Ember.RSVP.Promise((resolve, reject) => localAdapter.addHashForBulkUpdateOrCreate(localStore, modelType, snapshot, fieldsToUpdate, true)
+    .then(() => {
       let offlineGlobals = Ember.getOwner(_this).lookup('service:offline-globals');
       if (projection || (!projection && offlineGlobals.get('allowSyncDownRelatedRecordsWithoutProjection'))) {
         return syncDownRelatedRecords.call(_this, store, record, localAdapter, localStore, projection, params).then(() => {